requirements.txt
Kivy==2.1.0
Pillow==9.5.0
